requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.9.2
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
billiard==4.2.2
Brotli==1.1.0
celery==5.3.4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.10.7
cssselect2==0.8.0
defusedxml==0.7.1
Django==4.2.8
django-cors-headers==4.3.1
django-decouple==2.1
django-extensions==4.1
django-filter==23.5
django-picklefield==3.3
django-q2==1.8.0
djangorestframework==3.14.0
docxcompose==1.4.0
docxtpl==0.20.1
fonttools==4.60.0
frozenlist==1.7.0
gunicorn==21.2.0
html5lib==1.1
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
kombu==5.5.4
lxml==6.0.2
MarkupSafe==3.0.3
multidict==6.6.4
odfpy==1.4.1
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.9
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pypdf==6.1.2
pyphen==0.17.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.2.0
pytz==2025.2
redis==5.0.1
reportlab==4.4.4
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
twilio==9.8.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==60.2
webencodings==0.5.1
whitenoise==6.6.0
yarl==1.20.1
zopfli==0.2.3.post1

# Produção
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
